- Reworked ObjectExpression (and all Object*) to remove requirement for all function and mediator implementations to have to manage these types - Added user defined clr types to support scenarios where column data contains the representation of a more complex object; i.e. JSON, XML, binary serialized data - Reworked functions (only one required was Coalesce) to return a typed element when provided a generic type when fluently building the function (i.e. db.fx.Coalesce<int>(...) results in AnyElement<int> and db.fx.Coalesce(...) results in ObjectElement) - Reworked interfaces for *Elements to have a more correct inheritance hierarchy - Reworked fluent entry to select one and select many to accept more abstract type (i.e. AnyElement<bool> instead of BooleanElement) - Removed all typed select expressions in favor of generic select expression (i.e. SelectExpression<bool> instead of BooleanSelectExpression) - Hide TableAlias and FieldAlias properties from fluent building of query expressions - Reworked aliasing to return generic select expression (i.e. SelectExpression<bool> instead of BooleanSelectExpression)

+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/_Functions/CharIndex.cs
@@ -337,7 +337,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
             object result = exp.Execute();
 
             //then
-            result.Should().BeOfType<int>().Which.Should().Be(expected);
+            result.Should().BeOfType<long>().Which.Should().Be(expected);
         }
     }
 }
